+++ Irrigation/include/WebServer/WebServerException.h
@@ -4,12 +4,12 @@
 
 
 class WebServerException : public std::runtime_error {
-	unsigned statusCode;
-	KeyValue headers;
+	const unsigned statusCode;
+	const KeyValue headers;
 
 public:
-	WebServerException(unsigned statusCode, const std::string& errorMessage, const KeyValue& headers) :
-		std::runtime_error(errorMessage),
+	WebServerException(unsigned statusCode, const std::string& message = std::string(), const KeyValue& headers = KeyValue()) :
+		std::runtime_error(message),
 		statusCode(statusCode),
 		headers(headers)
 	{}
@@ -20,28 +20,14 @@ public:
 };
 
 
+class HttpNotAcceptable : public WebServerException {
+public:
+	HttpNotAcceptable() : WebServerException(406) {}
+};
+
+
 class HttpInternalServerError : public WebServerException {
 public:
-	HttpInternalServerError(const std::string& message, const KeyValue& headers = KeyValue()) : WebServerException(500, "Internal Server Error: " + message, headers) {}
-};
-
-class HttpNotFount : public WebServerException {
-public:
-	HttpNotFount(const std::string& message, const KeyValue& headers = KeyValue()) : WebServerException(404, "Not found: " + message, headers) {}
-};
-
-class HttpNotImplemented : public WebServerException {
-public:
-	HttpNotImplemented(const std::string& message, const KeyValue& headers = KeyValue()) : WebServerException(501, "Not implemented: " + message, headers) {}
-};
-
-class HttpNotSupported : public WebServerException {
-public:
-	HttpNotSupported(const std::string& message, const KeyValue& headers = KeyValue()) : WebServerException(505, "HTTP version is not supported: " + message, headers) {}
-};
-
-class HttpMethodNotAllowed : public WebServerException {
-public:
-	HttpMethodNotAllowed(const std::string& message, const KeyValue& headers = KeyValue()) : WebServerException(405, "Method not allowed: " + message, headers) {}
+	HttpInternalServerError() : WebServerException(500, "Internal Server Error", KeyValue({{"Content-Type", "text/plain"}})) {}
 };
