+++ Irrigation/test/Views/RestView/RestViewTest_programList.cpp
@@ -17,6 +17,7 @@ void RestViewTest::testGetProgramList(const list<ProgramDTO>& programDtoList) {
 	Response response = executeRequest("GET", createProgramListUrl());
 	EXPECT_THAT(response.responseCode, Eq(200));
 	EXPECT_THAT(response.writeCallbackData.text, Eq(XmlWriter().save(programDtoList)));
+	EXPECT_THAT(response.headerCallbackData.headers, Contains("Content-Type: application/xml\r\n"));
 }
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -33,13 +34,14 @@ TEST_F(RestViewTest, postProgramList) {
 	const IdType programId = document->getPrograms().begin()->first;
 
 	EXPECT_THAT(response.responseCode, Eq(201));
-	EXPECT_THAT(response.headerCallbackData.headers,
-			Contains("Location: /programs/" + to_string(programId) + "\r\n"));
+	EXPECT_THAT(response.headerCallbackData.headers, Contains("Location: /programs/" + to_string(programId) + "\r\n"));
+	EXPECT_THAT(response.headerCallbackData.headers, Not(Contains(HasSubstr("Content-Type:"))));
 }
 
 TEST_F(RestViewTest, postProgramListInvalidXml) {
 	const Response response = executeRequest("POST", createProgramListUrl(), "InvalidXml");
 	EXPECT_THAT(response.responseCode, Eq(400));
+	EXPECT_THAT(response.headerCallbackData.headers, Contains("Content-Type: application/xml\r\n"));
 }
 
 TEST_F(RestViewTest, getProgramList1) {
@@ -67,16 +69,19 @@ TEST_F(RestViewTest, getProgramList4) {
 }
 
 TEST_F(RestViewTest, putProgramList) {
-	Response response = executeRequest("PUT", createProgramListUrl());
+	Response response = executeRequest("PUT", createProgramListUrl(), XmlWriter().save(ProgramSample1().getDto()));
 	EXPECT_THAT(response.responseCode, Eq(405));
+	EXPECT_THAT(response.headerCallbackData.headers, Contains("Content-Type: application/xml\r\n"));
 }
 
 TEST_F(RestViewTest, patchProgramList) {
-	Response response = executeRequest("PATCH", createProgramListUrl());
+	Response response = executeRequest("PATCH", createProgramListUrl(), XmlWriter().save(ProgramSample1().getDto()));
 	EXPECT_THAT(response.responseCode, Eq(405));
+	EXPECT_THAT(response.headerCallbackData.headers, Contains("Content-Type: application/xml\r\n"));
 }
 
 TEST_F(RestViewTest, deleteProgramList) {
 	Response response = executeRequest("DELETE", createProgramListUrl());
 	EXPECT_THAT(response.responseCode, Eq(405));
+	EXPECT_THAT(response.headerCallbackData.headers, Contains("Content-Type: application/xml\r\n"));
 }
